Move `Workflow` class to the `wrapper` module
